move runtime macros into mod
